Move entry data property access and initialization into extension
